seperated test case into different files and few changes
