+++ rust/libnewsboat/tests/get_default_browser.rs
@@ -19,13 +19,3 @@ fn t_get_default_browser() {
     env::set_var(&key, &opera);
     assert_eq!(utils::get_default_browser(), opera);
 }
-
-#[test]
-fn t_resolve_tilde() {
-
-    env::set_var("HOME","test");    
-    assert_eq!(utils::resolve_tilde(String::from("~")),"test");
-    assert_eq!(utils::resolve_tilde(String::from("~/")),"test/");
-    assert_eq!(utils::resolve_tilde(String::from("~/foo/bar")),"test/foo/bar");
-    assert_eq!(utils::resolve_tilde(String::from("/foo/bar")),"/foo/bar");
-}
